More AST, up to Object Spec
Almost finished. I skipped the type spec options, but function and
object class spec are added. I'm not sure if I need to do param list or
some of the others. Most things should work at this point. I think it
would be advantageous to work on the generator now.