Added testCases.py
This will run test programs that are located in a directory. For
example the hard coded testDir is a relative path to the clone of
the Eidos examples provided by Boyanna[0]. If a file does not run
the program will continue and record the number of failed cases. You can
pipe the output to a file and search for WARNING to locate the specific
file that failed.
[0] https://github.com/CIS-UO/eidos-test