Bunch of error handling
Adding a bunch of error handling. Arithmetic operations will try try to
cast to the proper type, safely, and if they cannot then they will
generate an error telling the user the type of the values, or specific
value, that is causing the problem.

In general we should use the form
```
except Exception as e:
   raise Exception(e)
```
so that we can pass the exception along.