Now interpret in bash!!!
Added "eidos" program. This has two options to it. You can either enter
it with no arguments and you will enter a command line. This allows you
to run standard eidos type programs and then will return you the
resultant symbol table. You can also provide a file and it will run that
file.
In interpreter mode commands must be single line. But we do save the
previous result, so you can use a variable in a new line.
Files can be run with multiple lines of code.