IndexError handling for stack.pop
There are some "return result" statements there. I placed them under a
finally. Finally always runs after a try-catch block (it is part of it).
Is this how we want to handle? I'm not sure what your intended purpose
was here but I think this is right. Maybe we could reduce some of these
try blocks a little too. Need to discuss.