Hacky print
Commands like 'x;' will now just print the value of x, if it is in the
symbol library. Also moved the interpreter around a little bit. Moved
into just a python file. eidos file now calls `/usr/bin/env bash` which
appears to work on macs as well. But that file just calls the
interpreter with python3. So there's a fall back if that doesn't work.
Will have someone else test.