Error handling in interpreter
Added error handling for failing to enter a semicolon into the command
line. Yacc will return an error but we'll also get a print message that
tells us we can't do that. Also if you try to print ('x;') a variable
that does not exist, we will get a resulting error saying that and
return to our eidos prompt (returning the old symbol table). Point is to
maintain the old symbol table through errors and just tell the user they
fucked up, and not return to bash.