Regex added to interp
Added regex matching to variable names so that you can just output the
variable, as long as it is legal. Also added a 'display' function that
will list all stored variables. Mostly for debugging but can be useful.
Still unsure about how to handle functions since the value is a ast
node. Maybe we could get it to return something different or I could
have it call that stuff if it sees a node type. That might be the best
option.