Minor string changes
Type checked in addition and not in constant for string. This allows us
to add strings in the interpreter. There is still an error where you
can't pass a variable that is a string into a function, but you can pass
the string literals.
Eg: x="hello world"; f = foo(x);
fails
Eg: f = foo("Hello world");
works